+++ unit-testing-plugin/assets-map/index.js
@@ -1,0 +1,14 @@
+const path = require('path');
+
+module.exports.getChunksMap = (compilation) => {
+  const chunks = compilation.chunks;
+  const mappedChunks = chunks
+    .reduce((acc, c) => {
+      acc[c.name] = c.files
+      .filter((f) => path.extname(f) !== '.map');
+
+      return acc;
+    }, {});
+
+  return mappedChunks;
+};